refactor: move http and typeorm on module users to infra layer
