+++ src/modules/users/infra/http/controllers/UserAvatarController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import { instanceToInstance } from 'class-transformer';
-import { UpdateUserAvatarService } from '../services/UpdateUserAvatarService';
+import { UpdateUserAvatarService } from '../../../services/UpdateUserAvatarService';
 
 export class UserAvatarController {
   public async update(request: Request, response: Response): Promise<Response> {
