+++ src/modules/users/infra/http/controllers/UserController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { instanceToInstance } from 'class-transformer';
-import { ListUserService } from '../services/ListUserService';
-import { CreateUserService } from '../services/CreateUserService';
+import { ListUserService } from '../../../services/ListUserService';
+import { CreateUserService } from '../../../services/CreateUserService';
 
 export class UserController {
   public async index(request: Request, response: Response): Promise<Response> {
