+++ src/modules/users/infra/http/routes/profiles.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import { celebrate, Joi, Segments } from 'celebrate';
-import { AuthValidator } from '../../../shared/http/middlewares/AuthValidator';
+import { AuthValidator } from '../../../../../shared/infra/http/middlewares/AuthValidator';
 import { ProfileController } from '../controllers/ProfileController';
 
 function setupUserProfileRoutes(): Router {
